# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Montserrat Trails API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrailSource = typeof TrailSource[keyof typeof TrailSource];


export const TrailSource = {
  curated: 'curated',
  community: 'community',
} as const;

export interface Trail {
  id: number;
  name: string;
  distance_km: number;
  elevation_gain_m: number;
  difficulty: string;
  duration_hours: number;
  best_season: string;
  description: string;
  /** @nullable */
  terrain_type?: string | null;
  source: TrailSource;
  created_at: string;
}

export interface TrailInput {
  /** @minLength 1 */
  name: string;
  distance_km: number;
  elevation_gain_m: number;
  difficulty: string;
  duration_hours: number;
  best_season: string;
  description: string;
  terrain_type?: string;
}

export type TrailStatsByDifficultyItem = {
  difficulty: string;
  count: number;
};

export interface TrailStats {
  total_trails: number;
  curated_count: number;
  community_count: number;
  by_difficulty: TrailStatsByDifficultyItem[];
}

export interface ApiError {
  error: string;
}

export type ListTrailsParams = {
difficulty?: string;
min_duration_hours?: number;
max_duration_hours?: number;
terrain_type?: string;
};