# lib/api-zod/src/generated/types/trail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Montserrat Trails API
 * OpenAPI spec version: 0.1.0
 */
import type { TrailSource } from './trailSource';

export interface Trail {
  id: number;
  name: string;
  distance_km: number;
  elevation_gain_m: number;
  difficulty: string;
  duration_hours: number;
  best_season: string;
  description: string;
  /** @nullable */
  terrain_type?: string | null;
  source: TrailSource;
  created_at: string;
}