# lib/api-zod/src/generated/types/trailInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Montserrat Trails API
 * OpenAPI spec version: 0.1.0
 */

export interface TrailInput {
  /** @minLength 1 */
  name: string;
  distance_km: number;
  elevation_gain_m: number;
  difficulty: string;
  duration_hours: number;
  best_season: string;
  description: string;
  terrain_type?: string;
}