# lib/api-zod/src/generated/types/trailStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Montserrat Trails API
 * OpenAPI spec version: 0.1.0
 */
import type { TrailStatsByDifficultyItem } from './trailStatsByDifficultyItem';

export interface TrailStats {
  total_trails: number;
  curated_count: number;
  community_count: number;
  by_difficulty: TrailStatsByDifficultyItem[];
}